refs #41: restructure tests folder structure to match with core lib struct
