+++ composer.json
@@ -37,7 +37,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "LooplineSystems\\CloseIoApiWrapper\\Tests\\": "tests/LooplineSystems/CloseIoApiWrapper"
+            "Tests\\LooplineSystems\\CloseIoApiWrapper\\": "tests/LooplineSystems/CloseIoApiWrapper"
         }
     }
 }
