+++ tests/LooplineSystems/CloseIoApiWrapper/Api/CustomFieldApiTest.php
@@ -7,7 +7,7 @@
  * @license   https://github.com/loopline-systems/closeio-api-wrapper/blob/master/LICENSE (MIT Licence)
  */
 
-namespace LooplineSystems\CloseIoApiWrapper\Tests\Api;
+namespace Tests\LooplineSystems\CloseIoApiWrapper\Api;
 
 use LooplineSystems\CloseIoApiWrapper\Api\CustomFieldApi;
 use LooplineSystems\CloseIoApiWrapper\CloseIoApiWrapper;
