+++ tests/LooplineSystems/CloseIoApiWrapper/CloseIoApiWrapperTest.php
@@ -7,12 +7,12 @@
  * @license   https://github.com/loopline-systems/closeio-api-wrapper/blob/master/LICENSE (MIT Licence)
  */
 
-namespace LooplineSystems\CloseIoApiWrapper\Tests\Api;
+namespace Tests\LooplineSystems\CloseIoApiWrapper;
 
 use LooplineSystems\CloseIoApiWrapper\CloseIoApiWrapper;
 use LooplineSystems\CloseIoApiWrapper\CloseIoConfig;
 
-class WrapperTest extends \PHPUnit_Framework_TestCase
+class CloseIoApiWrapperTest extends \PHPUnit_Framework_TestCase
 {
     public function testCreateCloseIoWrapper()
     {
