+++ tests/LooplineSystems/CloseIoApiWrapper/CloseIoConfigTest.php
@@ -7,12 +7,12 @@
  * @license   https://github.com/loopline-systems/closeio-api-wrapper/blob/master/LICENSE (MIT Licence)
  */
 
-namespace LooplineSystems\CloseIoApiWrapper\Tests\Api;
+namespace Tests\LooplineSystems\CloseIoApiWrapper;
 
 use LooplineSystems\CloseIoApiWrapper\CloseIoConfig;
 use LooplineSystems\CloseIoApiWrapper\Library\Exception\InvalidParamException;
 
-class ConfigTest extends \PHPUnit_Framework_TestCase
+class CloseIoConfigTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @param string $url
