+++ tests/LooplineSystems/CloseIoApiWrapper/Library/Exception/ExceptionsTest.php
@@ -7,7 +7,7 @@
  * @license   https://github.com/loopline-systems/closeio-api-wrapper/blob/master/LICENSE (MIT Licence)
  */
 
-namespace LooplineSystems\CloseIoApiWrapper\Tests\Api;
+namespace Tests\LooplineSystems\CloseIoApiWrapper\Library\Exception;
 
 use LooplineSystems\CloseIoApiWrapper\Library\Exception\JsonDecodingException;
